+++ app/assets/scripts/main.js
@@ -4,8 +4,8 @@ var Router = require('react-router');
 var Route = Router.Route;
 var DefaultRoute = Router.DefaultRoute;
 
-var App = require('./components/app');
-var Home = require('./components/home');
+var App = require('./views/app');
+var Home = require('./views/home');
 
 var routes = (
   <Route handler={App}>
