+++ app/assets/scripts/views/app.js
